+++ lib/ooxml_parser/pptx_parser.rb
@@ -9,8 +9,9 @@ module OoxmlParser
     # @param path_to_file [String] file path
     # @return [Presentation] result of parse
     def self.parse_pptx(path_to_file)
-      Parser.parse_format(path_to_file) do |path_to_folder|
-        Presentation.new(unpacked_folder: path_to_folder).parse
+      file = OoxmlFile.new(path_to_file)
+      Parser.parse_format(file) do |yielded_file|
+        Presentation.new(unpacked_folder: yielded_file.path_to_folder).parse
       end
     end
   end